marshaler to codec and follow sdk simapp reorg
