+++ tests/e2e/ibc_fees_test.go
@@ -31,7 +31,7 @@ func TestIBCFeesTransfer(t *testing.T) {
 	//   with an ics-20 channel established
 	// when an ics-29 fee is attached to an ibc package
 	// then the relayer's payee is receiving the fee(s) on success
-	marshaler := app.MakeEncodingConfig(t).Marshaler
+	marshaler := app.MakeEncodingConfig(t).Codec
 	coord := wasmibctesting.NewCoordinator(t, 2)
 	chainA := coord.GetChain(wasmibctesting.GetChainID(1))
 	chainB := coord.GetChain(wasmibctesting.GetChainID(2))
@@ -116,7 +116,7 @@ func TestIBCFeesWasm(t *testing.T) {
 	//   and an ibc channel established
 	// when an ics-29 fee is attached to an ibc package
 	// then the relayer's payee is receiving the fee(s) on success
-	marshaler := app.MakeEncodingConfig(t).Marshaler
+	marshaler := app.MakeEncodingConfig(t).Codec
 	coord := wasmibctesting.NewCoordinator(t, 2)
 	chainA := coord.GetChain(wasmibctesting.GetChainID(1))
 	chainB := coord.GetChain(ibctesting.GetChainID(2))
